Fix gomod
