+++ tests/director/api/fixtures_queries_test.go
@@ -2,7 +2,10 @@ package api
 
 import (
 	"context"
+	"fmt"
 	"testing"
+
+	gcli "github.com/machinebox/graphql"
 
 	"github.com/kyma-incubator/compass/components/director/pkg/graphql"
 	"github.com/stretchr/testify/require"
@@ -314,4 +317,108 @@ func setDefaultEventingForApplication(t *testing.T, ctx context.Context, appID s
 	req := fixSetDefaultEventingForApplication(appID, runtimeID)
 	err := tc.RunOperation(ctx, req, nil)
 	require.NoError(t, err)
+}
+
+func createPackage(t *testing.T, ctx context.Context, appID, pkgName string) graphql.Package {
+	in, err := tc.graphqlizer.PackageCreateInputToGQL(fixPackageCreateInput(pkgName))
+	require.NoError(t, err)
+
+	req := gcli.NewRequest(
+		fmt.Sprintf(`mutation {
+			result: addPackage(applicationID: "%s", in: %s) {
+				id
+			}}`, appID, in))
+
+	var resp graphql.Package
+
+	err = tc.RunOperation(ctx, req, &resp)
+	require.NoError(t, err)
+
+	return resp
+}
+
+func getPackage(t *testing.T, ctx context.Context, appID, pkgID string) graphql.PackageExt {
+	req := fixPackageRequest(appID, pkgID)
+	pkg := graphql.ApplicationExt{}
+	require.NoError(t, tc.RunOperation(ctx, req, &pkg))
+	return pkg.Package
+}
+
+func deletePackage(t *testing.T, ctx context.Context, id string) {
+	req := gcli.NewRequest(
+		fmt.Sprintf(`mutation {
+			deletePackage(id: "%s") {
+				id
+			}}`, id))
+
+	require.NoError(t, tc.RunOperation(context.Background(), req, nil))
+}
+
+func addAPIToPackageWithInput(t *testing.T, ctx context.Context, pkgID string, input graphql.APIDefinitionInput) graphql.APIDefinitionExt {
+	inStr, err := tc.graphqlizer.APIDefinitionInputToGQL(input)
+	require.NoError(t, err)
+
+	actualApi := graphql.APIDefinitionExt{}
+	req := fixAddAPIToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &actualApi)
+	require.NoError(t, err)
+	return actualApi
+}
+
+func addAPIToPackage(t *testing.T, ctx context.Context, pkgID string) graphql.APIDefinitionExt {
+	input := fixAPIDefinitionInput()
+	inStr, err := tc.graphqlizer.APIDefinitionInputToGQL(input)
+	require.NoError(t, err)
+
+	actualApi := graphql.APIDefinitionExt{}
+	req := fixAddAPIToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &actualApi)
+	require.NoError(t, err)
+	return actualApi
+}
+
+func addEventToPackageWithInput(t *testing.T, ctx context.Context, pkgID string, input graphql.EventDefinitionInput) graphql.EventDefinition {
+	inStr, err := tc.graphqlizer.EventDefinitionInputToGQL(input)
+	require.NoError(t, err)
+
+	event := graphql.EventDefinition{}
+	req := fixAddEventAPIToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &event)
+	require.NoError(t, err)
+	return event
+}
+
+func addEventToPackage(t *testing.T, ctx context.Context, pkgID string) graphql.EventDefinition {
+	input := fixEventAPIDefinitionInput()
+	inStr, err := tc.graphqlizer.EventDefinitionInputToGQL(input)
+	require.NoError(t, err)
+
+	event := graphql.EventDefinition{}
+	req := fixAddEventAPIToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &event)
+	require.NoError(t, err)
+	return event
+}
+
+func addDocumentToPackageWithInput(t *testing.T, ctx context.Context, pkgID string, input graphql.DocumentInput) graphql.DocumentExt {
+	inStr, err := tc.graphqlizer.DocumentInputToGQL(&input)
+	require.NoError(t, err)
+
+	actualDoc := graphql.DocumentExt{}
+	req := fixAddDocumentToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &actualDoc)
+	require.NoError(t, err)
+	return actualDoc
+}
+
+func addDocumentToPackage(t *testing.T, ctx context.Context, pkgID string) graphql.DocumentExt {
+	input := fixDocumentInput()
+	inStr, err := tc.graphqlizer.DocumentInputToGQL(&input)
+	require.NoError(t, err)
+
+	actualDoc := graphql.DocumentExt{}
+	req := fixAddDocumentToPackageRequest(pkgID, inStr)
+	err = tc.RunOperation(ctx, req, &actualDoc)
+	require.NoError(t, err)
+	return actualDoc
 }
